+++ doc/changelog.rst
@@ -13,6 +13,9 @@ Version 0.2
 * Complete re-implementation of :ref:`fatiando.inversion <fatiando_inversion>`
   and all modules that depended on it. Inversion routines now have a standard
   interface. (`PR 72 <https://github.com/leouieda/fatiando/pull/72>`_)
+* Added moving window solution for Euler deconvolution in
+  :ref:`fatiando.gravmag.euler <fatiando_gravmag_euler>`.
+  (`PR 85 <https://github.com/leouieda/fatiando/pull/85>`_)
 * Renamed the ``fatiando.io`` module to
   :ref:`fatiando.datasets <fatiando_datasets>`
   (`PR 82 <https://github.com/leouieda/fatiando/pull/82>`_)